pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-coding-workflows"
version = "0.1.0"
description = "CLI scaffolding for AI-assisted coding workflows and a local conventional commit helper"
authors = [{ name = "EdisonTKPcom" }]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.0"
]

[project.scripts]
aiwf = "ai_workflows.cli:main"

[tool.setuptools.package-data]
ai_workflows = ["templates/**"]

[tool.setuptools.packages.find]
include = ["ai_workflows*"]
